brochure_generator/pyproject.toml
[project]
name = "brochure_generator"
version = "0.1.0"
description = "This project generates a brochure for a company to be used for prospective clients, investors and potential recruits.."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.43.0",
    "ipykernel>=6.30.1",
    "lxml>=6.0.2",
    "pydantic>=2.12.0",
    "python-decouple>=3.8",
    "streamlit>=1.50.0",
]
